Set dtype for neural ids to int32 (this caused problems before, since int was interpreted by cupy as either int32 or int64 depending on the version. Also fix minimal values of Np and Ne

+++ neuralflow/spike_data.py
@@ -123,7 +123,7 @@ class SpikeData:
                 self.cuda.cp.asarray(seq[0], dtype='float64')
             )
             self.cuda_var.data[iTrial].append(
-                self.cuda.cp.asarray(seq[1], dtype=int)
+                self.cuda.cp.asarray(seq[1], dtype='int32')
             )
 
     def trial_average_fr(self):
